requirements.txt
asgiref==3.9.0
crispy-bootstrap5==2025.6
diff-match-patch==20241021
dj-database-url==0.5.0
Django==4.2.23
django-allauth==65.10.0
django-crispy-forms==2.4
django-debug-toolbar==5.2.0
django-environ==0.12.0
gunicorn==20.1.0
pillow==11.3.0
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2==2.9.10
psycopg2-binary==2.9.10
setuptools==80.9.0
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.2
whitenoise==6.9.0
